add image preview. fixes #1
found that we can get a base64-encoded file body by adding
`?format=TEXT` to the end of the file path. with it it's fairly easy
to construct a dataURL and embed it in the page.

currently it only supports images (GIF, JPEG, PNG, SVG, and WebP);
not sure if any other binary formats are needed.

Issue: https://github.com/myakura/curcuma/issues/1